기능: Creator만 Video 접근, 수정, 삭제 가능

+++ controllers/videoController.js
@@ -54,7 +54,7 @@ export const videoDetail = async (req, res) => {
 
   try {
     const video = await Video.findById(id).populate("creator");
-    console.log(video);
+    // console.log(video);
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (err) {
     res.redirect(routes.home);
@@ -68,8 +68,13 @@ export const getEditVideo = async (req, res) => {
 
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (JSON.stringify(video.creator) === JSON.stringify(req.user.id)) {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    } else {
+      throw Error();
+    }
   } catch (err) {
+    console.log(err);
     res.redirect(routes.home);
   }
 };
@@ -93,7 +98,12 @@ export const deleteVideo = async (req, res) => {
   } = req;
 
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (JSON.stringify(video.creator) === JSON.stringify(req.user.id)) {
+      await Video.findOneAndRemove({ _id: id });
+    } else {
+      throw Error();
+    }
   } catch (err) {
     console.log(err);
   }
